Check for more warnings when compiling

+++ source/vpad_to_json.cpp
@@ -72,7 +72,7 @@ std::string pad_to_json(const PADData& pad_data)
     {
         writer.Key("wiiRemotes");
         writer.StartArray();
-        for(int i = 0; i < 4; ++i)
+        for(s8 i = 0; i < 4; ++i)
         {
             if(pad_data.wpad[i] == nullptr)
             {
@@ -186,7 +186,7 @@ std::string pad_to_json(const PADData& pad_data)
     {
         writer.Key("gameCubeControllers");
         writer.StartArray();
-        for(int i = 0; i < 4; ++i)
+        for(s8 i = 0; i < 4; ++i)
         {
             if(pad_data.pad[i] == nullptr)
             {
